Fixed some data type issues in fcnWDWAKE. #SaveTheBytes2K18

+++ fcnCREATEDSECT.m
@@ -4,11 +4,11 @@ function [sect] = fcnCREATEDSECT(sect, len, vars, first_idx, second_idx, first_p
 % vars is number of variables, 3 for VAP and 5 for HDVE
 
 rows = reshape(repmat([1:len]',1,vars)',[],1);
-col1 = reshape([repmat((first_idx.*vars)-(vars-1),1,vars) + repmat([0:(vars-1)], len,1)]',[],1);
+col1 = reshape([repmat((first_idx.*vars)-(vars-1),1,vars) + repmat(uint32([0:(vars-1)]), len,1)]',[],1);
 sect(sub2ind(size(sect),rows,col1)) = reshape(first_part',[],1);
 
 if ~isempty(second_idx)
-    col2 = reshape([repmat((second_idx.*vars)-(vars-1),1,vars) + repmat([0:(vars-1)], len,1)]',[],1);
+    col2 = reshape([repmat((second_idx.*vars)-(vars-1),1,vars) + repmat(uint32([0:(vars-1)]), len,1)]',[],1);
     sect(sub2ind(size(sect),rows,col2)) = reshape(second_part',[],1);
 end
 end
